\name{GetNenshuJP-package}
\alias{GetNenshuJP-package}
\alias{GetNenshuJP}
\docType{package}
\title{
Get data about Japanese company
}
\description{
This package provides Japanese company's data from www.nenshu.jp.
}
\details{
\tabular{ll}{
Package: \tab GetNenshuJP\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-10-20\cr
License: \tab MIT\cr
}
}
\author{
Hayakawa Atsushi

Maintainer: Hayakawa Atsushi <gepuro at mail.gepuro.net>

}
